move App.js to app dir
